refactor(web): 관리자 페이지 전수조사 — 항상 가짜였던 목업·자기참조·이원화 정리 + 관리자 전용 라우트 /admin 하위로
관리자 11 페이지 인벤토리에서 확인된 "실데이터처럼 보이는 가짜"와 구조 결함을 정리한다.

- analytics: 폐기된 7 brand profile(Pro/Think/Vision…) MODEL_USAGE·사인파 DAILY·고정 SUMMARY
  제거 — conv·latency 타일은 실 소스가 없어 항상 목업이었으므로 타일 자체를 뺌, 차트는 빈 상태
- metrics: 하드코딩 노드 6행(redis "degraded" 포함)·24h 사인파 요청률 차트·"—" 고정
  요청률/에러율 타일 제거(백엔드 시계열 없음)
- alerts: 규칙 5개 목업 + 동작 없는 [규칙 추가] + 로컬 전용 토글 제거 → 규칙은 코드 설정·
  시스템 설정(alerts 그룹)이 SoT 라는 안내 + 링크(알림 설정 이원화 해소)
- audit: audit_logs 에 컬럼이 없어 전부 'info' 고정이던 severity 필터·열 제거
- admin 대시보드: 자기 페이지를 가리키던 "사용자 관리" 퀵링크·"전체 보기" 제거
- 구조: AdminTabs 에만 있던 비-admin 경로 /mcp-monitoring·/agent-learning 을
  /admin/mcp-monitoring·/admin/agent-learning 으로 이동(구 라우트는 redirect 스텁)
- 고아 i18n 키 24건 4개 locale 동시 삭제(+ 기존 고아 admin.stats.totalUsersDelta)
