+++ apps/web/app/(workspace)/admin/analytics/page.tsx
@@ -25,45 +25,6 @@ const PERIODS: { key: Period; labelKey: string }[] = [
   { key: "90d", labelKey: "period.90d" },
 ];
 
-// 일별 대화량/모델 비중 실데이터: /api/metrics/analytics/daily-conversations,
-// /api/metrics/analytics/model-usage (conversation_messages 집계). 실패 시 아래 목업 유지.
-const DAILY: Record<Period, { label: string; value: number }[]> = {
-  "7d": [
-    { label: "weekday.1", value: 3120 },
-    { label: "weekday.2", value: 4280 },
-    { label: "weekday.3", value: 3890 },
-    { label: "weekday.4", value: 5210 },
-    { label: "weekday.5", value: 4820 },
-    { label: "weekday.6", value: 2310 },
-    { label: "weekday.0", value: 1980 },
-  ],
-  "30d": Array.from({ length: 30 }, (_, i) => ({
-    label: `${i + 1}`,
-    value: 2000 + Math.round(Math.abs(Math.sin(i / 3) * 4200)),
-  })),
-  "90d": Array.from({ length: 90 }, (_, i) => ({
-    label: `${i + 1}`,
-    value: 1800 + Math.round(Math.abs(Math.cos(i / 6) * 5200)),
-  })),
-};
-
-const SUMMARY: Record<Period, { conv: string; users: string; tokens: string; latency: string }> = {
-  "7d": { conv: "26,610", users: "318", tokens: "48.2M", latency: "1.12s" },
-  "30d": { conv: "112,940", users: "1,042", tokens: "214.7M", latency: "1.20s" },
-  "90d": { conv: "318,500", users: "1,510", tokens: "642.1M", latency: "1.18s" },
-};
-
-// 모델 프로파일별 사용 비중 목업 (실데이터 실패 시 폴백).
-const MODEL_USAGE = [
-  { name: "Default", pct: 34, color: "bg-m-default" },
-  { name: "Pro", pct: 21, color: "bg-m-pro" },
-  { name: "Fast", pct: 16, color: "bg-m-fast" },
-  { name: "Think", pct: 12, color: "bg-m-think" },
-  { name: "Code", pct: 9, color: "bg-m-code" },
-  { name: "Vision", pct: 5, color: "bg-m-vision" },
-  { name: "Auto", pct: 3, color: "bg-m-auto" },
-];
-
 /* 모델명 → 모델색 매핑 (usage 페이지와 동일 규칙) */
 const MODEL_BAR: Record<string, string> = {
   default: "bg-m-default",
@@ -117,10 +78,10 @@ function fmtCost(n: number) { return `$${n.toFixed(4)}`; }
 export default function AdminAnalyticsPage() {
   const t = useTranslations("adminAnalytics");
   const [period, setPeriod] = useState<Period>("7d");
-  // 실데이터 오버레이 (가능한 지표만): null 이면 목업 SUMMARY 사용
+  // 실데이터 오버레이 (가능한 지표만): null 이면 "—"
   const [liveUsers, setLiveUsers] = useState<string | null>(null);
   const [liveTokens, setLiveTokens] = useState<string | null>(null);
-  // 실데이터 차트: null 이면 목업 DAILY / MODEL_USAGE 사용
+  // 실데이터 차트: null 이면 빈 상태
   const [liveDaily, setLiveDaily] = useState<{ label: string; value: number }[] | null>(null);
   const [liveModels, setLiveModels] = useState<{ name: string; pct: number; color: string }[] | null>(null);
   // 추가 섹션 데이터
@@ -128,13 +89,10 @@ export default function AdminAnalyticsPage() {
   const [behaviorData, setBehaviorData] = useState<BehaviorData | null>(null);
   const [agentsData, setAgentsData] = useState<AgentRow[] | null>(null);
 
-  const daily =
-    liveDaily ??
-    (period === "7d"
-      ? DAILY["7d"].map((d) => ({ ...d, label: t(d.label) }))
-      : DAILY[period]);
-  const sum = SUMMARY[period];
-  const models = liveModels ?? MODEL_USAGE;
+  // 실데이터만 — 그전엔 사인파 DAILY·고정 SUMMARY·폐기된 7 brand profile(Pro/Think/Vision…)
+  // MODEL_USAGE 목업이 401/실패 시(그리고 conv·latency 는 항상) 그대로 화면에 나왔다.
+  const daily = useMemo(() => liveDaily ?? [], [liveDaily]);
+  const models = liveModels ?? [];
   const maxVal = useMemo(() => Math.max(1, ...daily.map((d) => d.value)), [daily]);
 
   useEffect(() => {
@@ -266,11 +224,9 @@ export default function AdminAnalyticsPage() {
 
       <div className="min-h-0 flex-1 overflow-y-auto p-6">
         <div className="grid grid-cols-1 gap-4 sm:grid-cols-2 lg:grid-cols-4">
-          {/* 값은 실데이터. 트렌드 델타는 실 비교 소스가 없어 제거(가짜 표시 방지). */}
-          <StatCard label={t("stats.totalConversations")} value={sum.conv} />
-          <StatCard label={t("stats.activeUsers")} value={liveUsers ?? sum.users} />
-          <StatCard label={t("stats.consumedTokens")} value={liveTokens ?? sum.tokens} />
-          <StatCard label={t("stats.avgResponse")} value={sum.latency} />
+          {/* 값은 실데이터만. 총 대화·평균 응답은 실 소스가 없어 타일 자체를 뺐다(항상 목업이었음). */}
+          <StatCard label={t("stats.activeUsers")} value={liveUsers ?? "—"} />
+          <StatCard label={t("stats.consumedTokens")} value={liveTokens ?? "—"} />
         </div>
 
         <div className="mt-6 grid grid-cols-1 gap-6 lg:grid-cols-3">
@@ -279,6 +235,7 @@ export default function AdminAnalyticsPage() {
               <CardTitle>{t("dailyConversations")}</CardTitle>
             </CardHeader>
             <CardContent>
+              {daily.length === 0 && <p className="py-10 text-center text-sm text-muted">{t("noData")}</p>}
               <div className="flex h-56 items-end gap-1.5">
                 {daily.map((d, i) => (
                   <div key={i} className="group flex h-full flex-1 flex-col items-center justify-end gap-1.5">
@@ -310,6 +267,7 @@ export default function AdminAnalyticsPage() {
               <CardTitle>{t("modelUsageTitle")}</CardTitle>
             </CardHeader>
             <CardContent>
+              {models.length === 0 && <p className="py-10 text-center text-sm text-muted">{t("noData")}</p>}
               <div className="space-y-3.5">
                 {models.map((m) => (
                   <div key={m.name}>
