+++ apps/web/app/(workspace)/admin/metrics/page.tsx
@@ -3,7 +3,7 @@
 import { useCallback, useEffect, useState } from "react";
 import { useLocale, useTranslations } from "next-intl";
 import { toBcp47 } from "@/i18n/config";
-import { Gauge, RefreshCw } from "lucide-react";
+import { RefreshCw } from "lucide-react";
 import {
   PageHeader,
   StatCard,
@@ -40,23 +40,6 @@ const STATUS_LABEL_KEY: Record<NodeStatus, string> = {
   degraded: "status.degraded",
   offline: "status.offline",
 };
-
-// 클러스터 노드/CPU/메모리는 /api/metrics·/api/metrics/metrics 실데이터로 오버레이.
-// 요청률 24h 시계열·요청률/에러율 StatCard 는 백엔드 시계열 미제공(usage-tracker stub)이라 목업 유지.
-const NODES: NodeRow[] = [
-  { name: "vllm-qwen-262k", role: "LLM · 262K", status: "online", latency: "0.82s", load: "61%" },
-  { name: "vllm-bge-m3", role: "Embedding", status: "online", latency: "0.04s", load: "23%" },
-  { name: "litellm-proxy", role: "Proxy :13401", status: "online", latency: "0.01s", load: "12%" },
-  { name: "postgres-primary", role: "Database", status: "online", latency: "0.6ms", load: "34%" },
-  { name: "redis-kv", role: "KV / Rate-limit", status: "degraded", latency: "1.2ms", load: "78%" },
-  { name: "api-server-2", role: "API (PM2)", status: "online", latency: "12ms", load: "44%" },
-];
-
-// 24시간 요청률 시계열 (req/min)
-const TIMESERIES = Array.from({ length: 24 }, (_, h) => ({
-  hour: h,
-  value: 120 + Math.round(Math.abs(Math.sin((h - 4) / 3) * 480) + (h >= 9 && h <= 18 ? 220 : 0)),
-}));
 
 type Cat = { id?: string; name?: string; status?: string; latency?: number | string };
 
@@ -165,7 +148,8 @@ export default function AdminMetricsPage() {
   const t = useTranslations("adminMetrics");
   const locale = toBcp47(useLocale());
   const [updated, setUpdated] = useState<string>("");
-  const [nodes, setNodes] = useState<NodeRow[]>(NODES);
+  // 클러스터 노드는 /api/metrics 실데이터만 — 그전엔 하드코딩 6행(redis "degraded" 포함)이 실렌더됐다.
+  const [nodes, setNodes] = useState<NodeRow[]>([]);
   const [cpu, setCpu] = useState<string | null>(null);
   const [mem, setMem] = useState<{ value: string; delta: string } | null>(null);
   const [agentSummary, setAgentSummary] = useState<AgentSummary | null>(null);
@@ -173,7 +157,6 @@ export default function AdminMetricsPage() {
   const [toolErrors, setToolErrors] = useState<ToolErrorData | null>(null);
   const [workflow, setWorkflow] = useState<WorkflowData | null>(null);
   const [routingGates, setRoutingGates] = useState<RoutingGatesData | null>(null);
-  const maxV = Math.max(...TIMESERIES.map((t) => t.value));
 
   const load = useCallback(async () => {
     try {
@@ -298,39 +281,11 @@ export default function AdminMetricsPage() {
 
       <div className="min-h-0 flex-1 overflow-y-auto p-6">
         <div className="grid grid-cols-1 gap-4 sm:grid-cols-2 lg:grid-cols-4">
-          {/* CPU·메모리는 실 시스템 지표. 요청률/에러율은 이 페이지에 실 소스가 없어
-              가짜 값(612/min·0.21%)+델타를 제거하고 "—"(데이터 없음)로 표시. */}
+          {/* CPU·메모리는 실 시스템 지표. 요청률/에러율 타일과 24h 요청률 차트는 백엔드 시계열이
+              없어 "—"·사인파 목업이었으므로 제거 — 소스가 생기면 그때 되살린다. */}
           <StatCard label={t("stat.cpu")} value={cpu ?? "—"} delta="load avg / cores" deltaTone="success" />
           <StatCard label={t("stat.memory")} value={mem?.value ?? "—"} delta={mem?.delta ?? t("memoryDelta", { total: "16" })} />
-          <StatCard label={t("stat.requestRate")} value="—" />
-          <StatCard label={t("stat.errorRate")} value="—" />
         </div>
-
-        <Card className="mt-6">
-          <CardHeader className="flex items-center gap-2">
-            <Gauge className="h-4 w-4 text-accent" />
-            <CardTitle>{t("requestChartTitle")}</CardTitle>
-          </CardHeader>
-          <CardContent>
-            <div className="flex h-48 items-end gap-1">
-              {TIMESERIES.map((bar) => (
-                <div
-                  key={bar.hour}
-                  className="group flex h-full flex-1 flex-col items-center justify-end gap-1"
-                  title={t("barTooltip", { hour: String(bar.hour).padStart(2, "0"), value: bar.value })}
-                >
-                  <div
-                    className="w-full rounded-t bg-accent/70 transition group-hover:bg-accent"
-                    style={{ height: `${Math.max(4, (bar.value / maxV) * 100)}%` }}
-                  />
-                  {bar.hour % 3 === 0 && (
-                    <span className="text-[9px] text-faint">{String(bar.hour).padStart(2, "0")}</span>
-                  )}
-                </div>
-              ))}
-            </div>
-          </CardContent>
-        </Card>
 
         <Card className="mt-6">
           <CardHeader>
@@ -348,6 +303,9 @@ export default function AdminMetricsPage() {
                 </tr>
               </thead>
               <tbody>
+                {nodes.length === 0 && (
+                  <tr><Td className="py-8 text-center text-muted" colSpan={5}>{t("noNodes")}</Td></tr>
+                )}
                 {nodes.map((n) => (
                   <tr key={n.name}>
                     <Td className="font-mono text-xs text-fg">{n.name}</Td>
