+++ apps/web/components/hub-tabs.tsx
@@ -68,8 +68,8 @@ export function AdminTabs() {
         { href: "/admin/conversations", label: tNav("items.conversationsAdmin") },
         { href: "/admin/analytics", label: tNav("items.analytics") },
         { href: "/admin/metrics", label: tNav("items.metrics") },
-        { href: "/mcp-monitoring", label: tNav("items.mcpMonitoring") },
-        { href: "/agent-learning", label: tNav("items.agentLearning") },
+        { href: "/admin/mcp-monitoring", label: tNav("items.mcpMonitoring") },
+        { href: "/admin/agent-learning", label: tNav("items.agentLearning") },
         { href: "/admin/agent-suggestions", label: tNav("items.agentSuggestionsAdmin") },
         { href: "/admin/audit", label: tNav("items.auditLog") },
         { href: "/admin/alerts", label: tNav("items.alerts") },
