+++ apps/web/messages/zh.json
@@ -1185,10 +1185,6 @@
       "guest": "访客"
     },
     "quickLinks": {
-      "users": {
-        "title": "用户管理",
-        "desc": "账户创建、权限与状态管理"
-      },
       "models": {
         "title": "模型 · MCP 设置",
         "desc": "模型预设与工具目录"
@@ -1200,7 +1196,6 @@
     },
     "stats": {
       "totalUsers": "总用户数",
-      "totalUsersDelta": "+12（过去 7 天）",
       "activeSessions": "活跃会话",
       "todayRequests": "今日请求",
       "systemStatus": "系统状态",
@@ -1243,10 +1238,8 @@
       "90d": "90 天"
     },
     "stats": {
-      "totalConversations": "总对话数",
       "activeUsers": "活跃用户",
-      "consumedTokens": "消耗令牌",
-      "avgResponse": "平均响应"
+      "consumedTokens": "消耗令牌"
     },
     "col": {
       "model": "模型",
@@ -1265,7 +1258,8 @@
       "4": "周四",
       "5": "周五",
       "6": "周六"
-    }
+    },
+    "noData": "没有可显示的数据。"
   },
   "adminConversations": {
     "title": "全部对话查看",
@@ -1325,12 +1319,8 @@
     "filter": {
       "allActions": "全部操作"
     },
-    "severity": {
-      "all": "全部严重级别"
-    },
     "status": {
       "critical": "严重",
-      "warn": "警告",
       "info": "信息"
     },
     "period": {
@@ -1343,8 +1333,7 @@
       "time": "时间",
       "actor": "操作者",
       "action": "操作",
-      "target": "目标",
-      "severity": "严重级别"
+      "target": "目标"
     },
     "empty": "没有符合条件的日志。"
   },
@@ -1355,13 +1344,9 @@
     "refresh": "刷新",
     "stat": {
       "cpu": "CPU 使用率",
-      "memory": "内存",
-      "requestRate": "请求率",
-      "errorRate": "错误率"
+      "memory": "内存"
     },
     "memoryDelta": "共 {total} GB 中",
-    "requestChartTitle": "请求率（最近 24 小时 · req/min）",
-    "barTooltip": "{hour}时: {value}/min",
     "nodeStatusTitle": "节点 · 服务状态",
     "nodeRoleLlm": "LLM 节点",
     "th": {
@@ -1434,14 +1419,13 @@
         "discussionHits": "讨论意图命中",
         "delegateHits": "委派意图命中"
       }
-    }
+    },
+    "noNodes": "没有上报的集群节点。"
   },
   "adminAlerts": {
     "title": "告警",
     "description": "管理告警规则并查看最近发生的事件。",
-    "addRule": "添加规则",
     "rulesTitle": "告警规则",
-    "toggleAria": "切换规则启用状态",
     "enabled": "已启用",
     "disabled": "已禁用",
     "recentTitle": "最近告警",
@@ -1452,29 +1436,9 @@
       "warning": "警告",
       "info": "信息"
     },
-    "rules": {
-      "cpu": {
-        "name": "CPU 使用率阈值",
-        "condition": "CPU > 85% · 持续 5 分钟"
-      },
-      "contextOverflow": {
-        "name": "LLM 上下文溢出",
-        "condition": "发生 ContextOverflowError"
-      },
-      "errorRate": {
-        "name": "错误率激增",
-        "condition": "5xx 比例 > 2%"
-      },
-      "tokenQuota": {
-        "name": "令牌配额耗尽",
-        "condition": "周配额超过 90%"
-      },
-      "roleChange": {
-        "name": "管理员权限变更",
-        "condition": "user.role_change → admin"
-      }
-    },
-    "empty": "暂无最近告警。"
+    "empty": "暂无最近告警。",
+    "rulesMovedHint": "告警规则（阈值、渠道）在代码配置和系统设置（alerts 组）中管理。此页面仅显示告警历史。",
+    "rulesMovedLink": "打开系统设置"
   },
   "adminMcpCatalog": {
     "close": "关闭",
